hard and soft delete

+++ controllers/classController.js
@@ -54,13 +54,26 @@ exports.updateClassById = async (req, res) => {
 
 exports.deleteClassById = async (req, res) => {
   const id = req.params.id;
+  const {harddelete} = req.body;
   try {
-    const classInstance = await Class.findByIdAndDelete(id);
+    if (harddelete) {
+    const deleted = await Class.findByIdAndDelete(id);
+      if (!deleted) {
+        return res.status(404).json({message:"class not found"});
+      }
+    res.status(200).json({ message: "Class deleted hardly successfully" });
+
+    }else{
+    const classInstance = await Class.findByIdAndUpdate(id,
+      {delete: true},
+{new:true}
+      );
     if (!classInstance) {
       return res.status(404).json({ error: "Class not found" });
     }
-    res.status(200).json({ message: "Class deleted successfully" });
-  } catch (error) {
+    res.status(200).json({ message: "Class deleted softly successfully" });
+  } 
+}catch (error) {
     res.status(400).json({ error: error.message });
   }
 };
